backend/felicity_lims/pyproject.toml
[tool.poetry]
name = "FelicityLims"
version = "0.1.0"
description = "FelicityLims"
authors = ["Aurthur Musendame <aurthurmusendame@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.6"
python-multipart = "^0.0.5"
emails = "^0.6"
email-validator = "^1.1.1"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
alembic = "^1.4.3"
sqlalchemy = "^1.3.19"
graphene = "^2.1.8"
databases = "^0.3.2"
asyncpg = "^0.21.0"
graphene-sqlalchemy = "^2.3.0"
sqlalchemy_mixins = "^1.2.1"
tenacity = "^6.2.0"

[tool.poetry.dev-dependencies]


